refactor: データベース接続の分離とモジュール化を実装
- ConnectionManagerクラスでコネクション管理を分離
- ヘルスチェック機能を独立したモジュールに分離
- 設定管理を専用ファイルに分離
- schema/index.tsでPrimarykeyとForeignKeyのバリデーション強化
- メインプロセスのデータベース初期化フローを簡素化
