+++ electron/main/index.ts
@@ -1,5 +1,6 @@
 import { app, BrowserWindow, shell, protocol } from 'electron'
 import { join } from 'node:path'
+import * as fs from 'node:fs'
 import { is } from '@electron-toolkit/utils'
 import { getDatabase, type DatabaseConfig } from './database/connection'
 
@@ -89,7 +90,7 @@ function createWindow(): void {
     callback({
       responseHeaders: {
         ...details.responseHeaders,
-        'Content-Security-Policy': is.dev 
+        'Content-Security-Policy': is.dev
           ? ['default-src \'self\' \'unsafe-inline\' \'unsafe-eval\' http://localhost:* ws://localhost:*; img-src \'self\' data: http://localhost:*;']
           : ['default-src \'self\' \'unsafe-inline\' \'unsafe-eval\'; style-src \'self\' \'unsafe-inline\'; img-src \'self\' data:; script-src \'self\' \'unsafe-inline\' \'unsafe-eval\';'],
       },
@@ -99,10 +100,10 @@ function createWindow(): void {
   // 環境判定の詳細ログ
   const isDevelopment = is.dev || process.env.NODE_ENV === 'development'
   const isProduction = !isDevelopment || process.env.NODE_ENV === 'production'
-  
+
   console.log(`🔍 環境判定: is.dev=${is.dev}, NODE_ENV=${process.env.NODE_ENV}`)
   console.log(`🔍 判定結果: isDevelopment=${isDevelopment}, isProduction=${isProduction}`)
-  
+
   // パフォーマンス最適化: 並列でリソースをプリロード
   if (isDevelopment) {
     // Development: Connect to Nuxt dev server
@@ -126,8 +127,8 @@ function createWindow(): void {
     // Production: Load built Nuxt files - 200.htmlを使用してSPAルーティングに対応
     const htmlPath = join(__dirname, '../../.output/public/200.html')
     console.log(`🔍 プロダクション用HTMLファイルパス: ${htmlPath}`)
-    console.log(`📂 HTMLファイルの存在確認: ${require('fs').existsSync(htmlPath)}`)
-    
+    console.log(`📂 HTMLファイルの存在確認: ${fs.existsSync(htmlPath)}`)
+
     mainWindow.loadFile(htmlPath)
 
     // プロダクション時の最適化とデバッグログ
@@ -172,14 +173,14 @@ if (!is.dev) {
  */
 async function initializeDatabase(): Promise<void> {
   const dbStartTime = Date.now()
-  
+
   try {
     const db = getDatabase()
-    
+
     // 環境判定
     const isDevelopment = is.dev || process.env.NODE_ENV === 'development'
     const isTest = process.env.NODE_ENV === 'test'
-    
+
     // データベース設定
     const dbConfig: Partial<DatabaseConfig> = {
       environment: isTest ? 'test' : isDevelopment ? 'development' : 'production',
@@ -191,20 +192,20 @@ async function initializeDatabase(): Promise<void> {
       enableMigrations: true,
       maxConnections: isDevelopment ? 3 : 5,
       connectionTimeout: 30000,
-      enableLogging: isDevelopment && process.env.DATABASE_DEBUG === 'true'
+      enableLogging: isDevelopment && process.env.DATABASE_DEBUG === 'true',
     }
-    
+
     console.log(`🗄️ データベース初期化開始 (環境: ${dbConfig.environment})`)
-    
+
     // データベース初期化とマイグレーション実行
     await db.initialize(dbConfig)
-    
+
     // 接続テスト
     const isHealthy = await db.testConnection()
     if (!isHealthy) {
       throw new Error('データベース接続テストに失敗しました')
     }
-    
+
     // ヘルスチェック（本番環境のみ）
     if (!isDevelopment) {
       const health = await db.healthCheck()
@@ -212,41 +213,41 @@ async function initializeDatabase(): Promise<void> {
         console.warn('⚠️ データベースヘルスチェックで問題が検出されました:', health.issues)
       }
     }
-    
+
     databaseInitTime = Date.now() - dbStartTime
     console.log(`✅ データベース初期化完了: ${databaseInitTime}ms`)
-    
+
     // パフォーマンス監視用ログ
     if (process.env.PERFORMANCE_MONITOR) {
       const status = db.getStatus()
       console.log('Database ready', {
         initTime: databaseInitTime,
         environment: status.environment,
-        isHealthy: status.connectionInfo?.isConnected || false
+        isHealthy: status.connectionInfo?.isConnected || false,
       })
     }
-    
-  } catch (error) {
+  }
+  catch (error) {
     databaseInitTime = Date.now() - dbStartTime
     const errorMessage = error instanceof Error ? error.message : String(error)
-    
+
     // 環境判定（エラーハンドリング用）
     const isDev = is.dev || process.env.NODE_ENV === 'development'
-    
+
     console.error(`❌ データベース初期化に失敗しました (${databaseInitTime}ms):`, errorMessage)
-    
+
     // 開発時はエラーの詳細を表示
     if (isDev) {
       console.error('データベースエラーの詳細:', error)
     }
-    
-    
+
     // データベース初期化エラーは致命的なため、アプリケーションを終了
     // ただし、開発時は警告のみ表示してアプリケーションを継続
     if (!isDev) {
       app.quit()
       return
-    } else {
+    }
+    else {
       console.warn('⚠️ 開発環境のため、データベースエラーを無視してアプリケーションを継続します')
     }
   }
@@ -259,7 +260,7 @@ app.whenReady().then(async () => {
   // パフォーマンス測定
   const appReadyTime = Date.now() - startTime
   console.log(`⚡ Electronアプリ初期化完了: ${appReadyTime}ms`)
-  
+
   // データベース初期化（並列実行）
   const databaseInitPromise = initializeDatabase()
 
@@ -277,7 +278,8 @@ app.whenReady().then(async () => {
   // データベース初期化の完了を待つ
   try {
     await databaseInitPromise
-  } catch (error) {
+  }
+  catch (error) {
     console.error('データベース初期化エラー:', error)
   }
 
@@ -291,13 +293,15 @@ app.whenReady().then(async () => {
           try {
             window.webContents.openDevTools({ mode: 'detach' })
             console.log('🔧 開発時DevToolsを開きました')
-          } catch (error) {
+          }
+          catch (error) {
             console.error('❌ DevToolsを開けませんでした:', error)
             // 代替方法でDevToolsを開く
             try {
               window.webContents.toggleDevTools()
               console.log('🔧 代替方法でDevToolsを開きました')
-            } catch (altError) {
+            }
+            catch (altError) {
               console.error('❌ 代替方法でもDevToolsを開けませんでした:', altError)
             }
           }
@@ -318,8 +322,8 @@ app.whenReady().then(async () => {
           window.webContents.toggleDevTools()
         }
         // Ctrl+Shift+I (Windows/Linux) または Cmd+Opt+I (Mac) でDevToolsの切り替え
-        if (input.key === 'I' && input.type === 'keyDown' && 
-            ((input.control && input.shift) || (input.meta && input.alt))) {
+        if (input.key === 'I' && input.type === 'keyDown'
+          && ((input.control && input.shift) || (input.meta && input.alt))) {
           window.webContents.toggleDevTools()
         }
       })
@@ -330,9 +334,9 @@ app.whenReady().then(async () => {
       // リソース読み込み状況をデバッグ
       window.webContents.on('did-finish-load', () => {
         console.log('🎯 ページの読み込みが完了しました')
-      
-      // JavaScriptエラーをキャッチ
-      window.webContents.executeJavaScript(`
+
+        // JavaScriptエラーをキャッチ
+        window.webContents.executeJavaScript(`
         window.addEventListener('error', (e) => {
           console.error('🚨 JS Error:', e.error, 'File:', e.filename, 'Line:', e.lineno);
         });
@@ -377,14 +381,14 @@ app.whenReady().then(async () => {
         }, 5000);
       `).catch(err => console.error('JavaScriptの実行に失敗:', err))
       })
-      
+
       // リソース読み込みエラーをキャッチ
       window.webContents.on('did-fail-load', (_, errorCode, errorDescription, validatedURL, isMainFrame) => {
         console.error('📛 リソース読み込み失敗:', {
           errorCode,
           errorDescription,
           validatedURL,
-          isMainFrame
+          isMainFrame,
         })
       })
 
@@ -444,7 +448,8 @@ app.on('before-quit', async () => {
     const db = getDatabase()
     await db.cleanup()
     console.log('🗄️ データベース接続をクリーンアップしました')
-  } catch (error) {
+  }
+  catch (error) {
     console.error('データベースクリーンアップエラー:', error)
   }
 
@@ -459,7 +464,7 @@ if (is.dev) {
   setInterval(() => {
     const memUsage = process.memoryUsage()
     console.log(`🧠 メモリ使用量: RSS=${Math.round(memUsage.rss / 1024 / 1024)}MB, Heap=${Math.round(memUsage.heapUsed / 1024 / 1024)}MB`)
-    
+
     // データベースのパフォーマンス統計も表示
     try {
       const db = getDatabase()
@@ -467,7 +472,8 @@ if (is.dev) {
       if (status.isInitialized && status.performance) {
         console.log(`🗄️ DB統計: クエリ数=${status.performance.queryCount}, 平均時間=${status.performance.averageQueryTime.toFixed(2)}ms, 低速クエリ=${status.performance.slowQueryCount}`)
       }
-    } catch (error) {
+    }
+    catch {
       // データベースが初期化されていない場合は無視
     }
   }, 30000) // 30秒ごと
